test(kernel): Add comprehensive tests for message pipe IPC
Add 20 new tests covering pipe_init, pipe_enqueue, and pipe_dequeue:
- pipe_init: basic, invalid index, zero capacity, max capacity, too large, already engaged, multiple, boundary
- pipe_enqueue: invalid index, not engaged, message too large, basic, multiple
- pipe_dequeue: invalid index, not engaged, message too large, basic
- Integration: enqueue/dequeue together, circular buffer wrap-around

Also fixed:
- __init_pipe loop variable: uint8_t -> uint16_t (MAX_MESSAGE_BUFFER_BYTES is 128)
- Updated semaphore tests to use max_count instead of init_count

All 102 tests pass.

+++ Core/Src/kernel/task.c
@@ -210,10 +210,10 @@ static inline void __init_sem(uint8_t semaphore_idx, uint32_t semaphore_count, b
 static inline void __init_pipe(uint8_t message_pipe_idx, uint8_t max_messages, bool should_engage) {
     message_pipe_list[message_pipe_idx]->count = 0;
     message_pipe_list[message_pipe_idx]->max_count = max_messages;
-    for (uint8_t i = 0; i < MAX_MESSAGE_BUFFER_BYTES; i++) {
+    for (uint16_t i = 0; i < MAX_MESSAGE_BUFFER_BYTES; i++) {
         message_pipe_list[message_pipe_idx]->buffer[i] = 0;
     }
-    message_pipe_list[message_pipe_idx]->enqueue_idx= 0;
+    message_pipe_list[message_pipe_idx]->enqueue_idx = 0;
     message_pipe_list[message_pipe_idx]->dequeue_idx = 0;
     message_pipe_list[message_pipe_idx]->tick_updated_at = os_tick_count;
     message_pipe_list[message_pipe_idx]->engaged = should_engage;
